requirements.txt
fastapi==0.115.0
pydantic==2.9.2
uvicorn==0.31.1
websockets==13.1
sqlalchemy==2.0.35
alembic==1.13.3
bcrypt==4.0.1
passlib==1.7.4
python-jose==3.3.0
aiosqlite==0.20.0
pydantic_settings==2.5.2
jinja2==3.1.4
python-multipart==0.0.5
email-validator==2.2.0